requirements.txt
pandas>=1.3.0
sentence-transformers>=2.0.0
faiss-cpu>=1.7.0
streamlit
